+++ R/dbi-table.R
@@ -126,7 +126,7 @@ setMethod("dbAppendTable", "OdbcConnection",
     stopifnot(is.null(row.names))
 
     fieldDetails <- tryCatch({
-      details <- odbcConnectionColumns(conn, name, exact = TRUE)
+      details <- odbcConnectionColumns_(conn, name, exact = TRUE)
       details$param_index <- match(details$name, colnames(value))
       details[!is.na(details$param_index) & !is.na(details$data_type), ]
     },