+++ NEWS.md
@@ -1,5 +1,8 @@
 # odbc (development version)
 
+* Deprecated `odbcConnectionColumns()` (in favor of `dbListFields()`), 
+  `odbcConnectionActions()`, and `odbcConnectionIcon()` (@simonpcouch, #699).
+  
 * `dbConnect()` no longer automatically escapes suspicious characters
   (since there doesn't seem to be a consistent way to do this across drivers)
   but instead points you to `quote_value()` which applies a heuristic